pyproject.toml
[project]
name = "ecommerce-smart-filtering"
version = "0.1.0"
description = "A smart filtering system for e-commerce platforms"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi[standard]>=0.115.12",
    "openai>=1.73.0",
    "pillow>=11.1.0",
    "playwright>=1.51.0",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "structlog>=25.2.0",
]

[project.optional-dependencies]
local = [
    "accelerate>=1.6.0",
    "datasets>=3.5.0",
    "outlines>=0.2.3",
    "sentencepiece>=0.2.0",
    "transformers>=4.51.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.5",
]
